Bug fixing roundoff error in distance
Unfortunately we overestimate now....

+++ kernel/train/Poly.h
@@ -10,7 +10,7 @@ typedef struct Poly {
 
 void poly_init(Poly* p, int t0, int t1, int v0, int v1);
 
-int eval_dist(const Poly const* p, int t);
+float eval_dist(const Poly const* p, int t);
 int eval_velo(const Poly const* p, int t);
 
 #endif // Poly.h
